refactor manifest: rewrite manifest update fns as methods
