+++ src/supertable/manifest/commit.rs
@@ -41,21 +41,15 @@
 //! URI predictable before any PUT); a serial implementation
 //! is correctness-equivalent but pessimistic on object stores.
 
-use std::collections::{BTreeMap, HashMap, HashSet};
 use std::sync::Arc;
-
-use futures::future;
 
 use crate::storage::{ObjectMeta, StorageError, StorageProvider};
 use crate::supertable::error::CommitError;
-use crate::supertable::manifest::list::{
-    self as list_mod, ManifestList, ManifestListEntry, PartitionStrategy,
-};
+use crate::supertable::manifest::list::{self as list_mod, ManifestList};
 use crate::supertable::manifest::part::{
     self as part_mod, BLAKE3_DIGEST_BYTES, BLAKE3_HEX_LEN, ContentHash, ManifestPart, PartId,
 };
-use crate::supertable::manifest::partition::{assign_partition, encode_partition_key};
-use crate::supertable::{Manifest, ManifestLoadError, SuperfileEntry, SupertableOptions};
+use crate::supertable::{Manifest, ManifestLoadError};
 
 /// Pointer-file location under the supertable root. The only
 /// path that ever gets atomically renamed; everything else is
@@ -68,12 +62,6 @@ pub const MANIFEST_LISTS_DIR: &str = "manifest-lists";
 
 /// Subdirectory for manifest part files.
 pub const MANIFEST_PARTS_DIR: &str = "manifests";
-
-/// Zstd compression level for manifest parts and the manifest list.
-/// Level 3 is zstd's own default — a balanced ratio/speed point that
-/// keeps commit latency low while compressing the Avro-encoded
-/// manifest well. (Valid range is 1..=22.)
-pub const MANIFEST_ZSTD_LEVEL: i32 = 3;
 
 /// Build the URI for a manifest list at a given manifest_id.
 /// 6-digit zero-pad gives stable lexicographic ordering for
@@ -183,6 +171,10 @@ impl PointerFile {
                 .ok_or_else(|| ManifestLoadError::PointerParse("missing content_hash".into()))?,
         })
     }
+
+    pub fn get_manifest_id(&self) -> u64 {
+        self.manifest_id
+    }
 }
 
 /// Read the pointer file from storage.
@@ -264,7 +256,7 @@ pub async fn write_manifest_part(
 
 /// Read the decompressed size from the zstd frame header in O(1), no
 /// actual decompression.
-fn frame_content_size(compressed: &[u8], fallback: u64) -> u64 {
+pub(crate) fn frame_content_size(compressed: &[u8], fallback: u64) -> u64 {
     zstd::zstd_safe::get_frame_content_size(compressed)
         .ok()
         .flatten()
@@ -272,7 +264,7 @@ fn frame_content_size(compressed: &[u8], fallback: u64) -> u64 {
 }
 
 /// PUT pre-encoded part bytes to storage.
-async fn write_part_bytes(
+pub(crate) async fn write_part_bytes(
     storage: &dyn StorageProvider,
     encoded: &[u8],
 ) -> Result<(), CommitError> {
@@ -339,70 +331,6 @@ pub async fn write_pointer(
     }
 }
 
-/// Commit a new manifest version.
-///
-/// Orchestrates the four-step sequence:
-///
-/// 1. **In parallel** — write each new manifest part + write
-///    the new manifest list. Independent of each other; the
-///    list references parts by URI (= blake3 of bytes,
-///    computed before any I/O). Issued via
-///    [`futures::future::join_all`].
-/// 2. Await all of the above (visibility barrier #1: parts
-///    and list must be durable before the pointer publishes).
-/// 3. Build the new pointer file (manifest_id, list_uri,
-///    list_content_hash).
-/// 4. Conditional pointer-PUT (visibility barrier #2: the
-///    rename is the only thing readers observe).
-///
-/// `parts_to_write` should contain **only the parts that need
-/// to be persisted** (i.e., new + changed). Each element is the
-/// pre-encoded (Avro+zstd) bytes produced by [`build_part_and_entry`]
-/// — passing them directly avoids a second encode cycle.
-/// Reused parts from the previous manifest version are not in this
-/// list — their URIs are already in `new_list.parts[i].uri`.
-pub async fn commit_manifest(
-    storage: &dyn StorageProvider,
-    expected_prev_etag: Option<&str>,
-    new_list: &ManifestList,
-    parts_to_write: &[&[u8]],
-) -> Result<PointerFile, CommitError> {
-    // Step 1+2: parallel write of (list, parts).
-    //
-    // Both futures are independent — the list's references to
-    // each part's URI are content-addressable from the
-    // in-memory bytes before any I/O, so there's no
-    // happens-before edge between them.
-    let list_fut = write_manifest_list(storage, new_list);
-    let part_futs = parts_to_write
-        .iter()
-        .map(|encoded| write_part_bytes(storage, encoded));
-    let part_join = future::join_all(part_futs);
-
-    let (list_res, part_results) = tokio::join!(list_fut, part_join);
-    // Translate `Storage(PreconditionFailed)` from sub-writes
-    // into `WriteContentionExhausted` so callers (and the
-    // writer's OCC retry loop) can match on one variant
-    // regardless of which CAS lost the race — list or pointer.
-    let list_res = list_res.map_err(translate_contention)?;
-    for part_result in part_results {
-        part_result.map_err(translate_contention)?;
-    }
-
-    // Step 3: build pointer.
-    let pointer = PointerFile {
-        manifest_id: new_list.manifest_id,
-        manifest_list_uri: list_res.uri,
-        content_hash: list_res.content_hash,
-    };
-
-    // Step 4: conditional pointer write — the visibility
-    // barrier. Until this succeeds, no reader sees the new
-    // manifest version.
-    write_pointer(storage, &pointer, expected_prev_etag).await?;
-    Ok(pointer)
-}
-
 /// Test-helper alias so test code can construct a
 /// `Arc<dyn StorageProvider>` and pass it through this
 /// module's `&dyn StorageProvider`-typed APIs in one cast.
@@ -417,7 +345,7 @@ pub fn as_dyn(p: &Arc<dyn StorageProvider>) -> &dyn StorageProvider {
 /// same manifest_id. Caller maps to OCC retry or to a
 /// terminal "write contention" error to the user. Other
 /// errors pass through unchanged.
-fn translate_contention(e: CommitError) -> CommitError {
+pub(crate) fn translate_contention(e: CommitError) -> CommitError {
     match e {
         CommitError::Storage(StorageError::PreconditionFailed { .. }) => {
             CommitError::WriteContentionExhausted
@@ -433,266 +361,20 @@ pub async fn get_current_manifest_etag(
     storage: &Arc<dyn StorageProvider>,
     current: Arc<Manifest>,
 ) -> Result<Option<String>, CommitError> {
-    let Some((pointer_file, meta)) = read_pointer(storage.as_ref()).await? else {
+    let Some((pointer_file, meta)) = read_pointer(storage.as_ref())
+        .await
+        .map_err(|e| CommitError::PointerParse(e.to_string()))?
+    else {
         return Ok(None);
     };
     let Some(meta_list) = current.list.as_ref() else {
         // no manifest list for in-memory supertables
         return Ok(None);
     };
-    if pointer_file.manifest_id == meta_list.manifest_id {
+    if pointer_file.get_manifest_id() == meta_list.manifest_id {
         return Ok(meta.etag);
     }
     Err(CommitError::WriteContentionExhausted)
-}
-
-/// build one `ManifestPart` from `superfiles` + the
-/// matching `ManifestListEntry`. Encodes the part once,
-/// content-hashes it, and computes the list-level aggregate
-/// skip summaries that `list_prune` reads at query time.
-fn build_part_and_entry(
-    opts: &SupertableOptions,
-    superfiles: Vec<Arc<SuperfileEntry>>,
-    partition_key: Vec<u8>,
-) -> Result<
-    (
-        crate::supertable::manifest::list::ManifestListEntry,
-        crate::supertable::manifest::part::ManifestPart,
-        Vec<u8>, // pre-encoded compressed bytes — reused by write path, no second encode
-    ),
-    crate::supertable::CommitError,
-> {
-    let _ = opts; // reserved for future per-options encoding tweaks (zstd level, etc.)
-
-    let part = ManifestPart {
-        format_version: part_mod::FORMAT_VERSION.into(),
-        part_id: PartId::new_v4(),
-        superfiles,
-    };
-    let compressed = part_mod::encode(&part, MANIFEST_ZSTD_LEVEL);
-    let size_compressed = compressed.len() as u64;
-    let content_hash = ContentHash::of(&compressed);
-    let size_uncompressed = frame_content_size(&compressed, size_compressed);
-    let aggregates = crate::supertable::manifest::aggregates::compute(&part.superfiles);
-    let entry = ManifestListEntry {
-        part_id: part.part_id,
-        uri: part_uri(&content_hash),
-        n_superfiles: part.superfiles.len() as u64,
-        size_bytes_compressed: size_compressed,
-        size_bytes_uncompressed: size_uncompressed,
-        content_hash,
-        partition_key,
-        id_range: aggregates.id_range,
-        scalar_stats_agg: aggregates.scalar_stats_agg,
-        fts_summary_agg: aggregates.fts_summary_agg,
-        vector_summary_agg: aggregates.vector_summary_agg,
-    };
-    Ok((entry, part, compressed))
-}
-
-/// Returns the new ManifestListEntries when `new_entries` are added to `old` manifest. This
-/// operation may create new ManifestParts. The function also returns the new ManifestParts that
-/// the caller can decide to write to storage.
-pub async fn rebalance_for_commit(
-    opts: &Arc<SupertableOptions>,
-    old: &Arc<crate::supertable::Manifest>,
-    new_entries: &[Arc<SuperfileEntry>],
-    entries_to_remove: &[Arc<SuperfileEntry>],
-) -> Result<(Vec<ManifestListEntry>, Vec<EncodedPart>), CommitError> {
-    // 1. Resolve the effective partition strategy. Locked at
-    //    first commit: read from the existing manifest list
-    //    if present, else use the options default.
-    let strategy: PartitionStrategy = old
-        .list
-        .as_ref()
-        .map(|l| l.partition_strategy.clone())
-        .unwrap_or_else(|| opts.effective_partition_strategy());
-
-    // 2. Group new entries by partition_key (the on-disk
-    //    encoding the list + parts carry).
-    let mut new_by_partition: BTreeMap<Vec<u8>, Vec<Arc<SuperfileEntry>>> = BTreeMap::new();
-    for entry in new_entries {
-        let pk = assign_partition(entry, &strategy)?;
-        new_by_partition
-            .entry(encode_partition_key(&pk))
-            .or_default()
-            .push(Arc::clone(entry));
-    }
-
-    let mut removals_by_partition: BTreeMap<Vec<u8>, Vec<Arc<SuperfileEntry>>> = BTreeMap::new();
-    for entry in entries_to_remove {
-        let pk = assign_partition(entry, &strategy)?;
-        removals_by_partition
-            .entry(encode_partition_key(&pk))
-            .or_default()
-            .push(Arc::clone(entry));
-    }
-
-    // 3. Walk the existing list entries, classify each by
-    //    whether it's the *latest* entry for its partition.
-    //    The "rewrite latest part" policy: only the
-    //    most recent entry per partition is a candidate for
-    //    rewrite; older entries for the same partition (from
-    //    a prior part-split) carry over unchanged.
-    let mut latest_index_for_partition: HashMap<Vec<u8>, usize> = HashMap::new();
-    if let Some(old_list) = old.list.as_ref() {
-        for (i, entry) in old_list.parts.iter().enumerate() {
-            latest_index_for_partition.insert(entry.partition_key.clone(), i);
-        }
-    }
-    // The output list entries — built incrementally as we
-    // walk existing entries + emit new ones for cold
-    // partitions. Order: existing entries (touched ones
-    // replaced in place; untouched preserved) followed by
-    // entries for cold partitions.
-    let mut out_list_entries: Vec<ManifestListEntry> = Vec::new();
-    let mut parts_to_write: Vec<EncodedPart> = Vec::new();
-    let mut handled_partitions: HashSet<Vec<u8>> = HashSet::new();
-
-    if let Some(old_list) = old.list.as_ref() {
-        for (i, entry) in old_list.parts.iter().enumerate() {
-            let is_latest_for_partition = latest_index_for_partition
-                .get(&entry.partition_key)
-                .copied()
-                == Some(i);
-            let touched = new_by_partition.contains_key(&entry.partition_key);
-
-            if is_latest_for_partition && touched {
-                let new_for_pk = new_by_partition
-                    .remove(&entry.partition_key)
-                    .expect("touched implies present");
-
-                let combined_n = entry.n_superfiles as usize + new_for_pk.len();
-                if combined_n as u64 > opts.target_superfiles_per_partition {
-                    // Split: keep the existing entry as-is and emit a
-                    // fresh part with just the new superfiles.
-                    out_list_entries.push(entry.clone());
-                    let (fresh_entry, fresh_part, fresh_encoded) =
-                        build_part_and_entry(opts, new_for_pk, entry.partition_key.clone())?;
-                    out_list_entries.push(fresh_entry);
-                    parts_to_write.push(EncodedPart {
-                        part: fresh_part,
-                        encoded: fresh_encoded,
-                    });
-                } else {
-                    // Rewrite: load existing part and combine with new superfiles.
-                    let existing_part = old.get_part_by_id(entry.part_id).await.map_err(|e| {
-                        crate::supertable::CommitError::PointerParse(format!(
-                            "loading existing part {} for partition rewrite: {e}",
-                            entry.part_id.0
-                        ))
-                    })?;
-                    let combined_superfiles: Vec<Arc<SuperfileEntry>> = existing_part
-                        .superfiles
-                        .iter()
-                        .cloned()
-                        .chain(new_for_pk)
-                        .collect();
-                    let (rebuilt_entry, rebuilt_part, rebuilt_encoded) = build_part_and_entry(
-                        opts,
-                        combined_superfiles,
-                        entry.partition_key.clone(),
-                    )?;
-                    out_list_entries.push(rebuilt_entry);
-                    parts_to_write.push(EncodedPart {
-                        part: rebuilt_part,
-                        encoded: rebuilt_encoded,
-                    });
-                }
-                handled_partitions.insert(entry.partition_key.clone());
-            } else {
-                // Carry over: either an older entry for a
-                // touched partition (handled when we hit the
-                // latest), or an entry for an untouched
-                // partition. Either way, content-hash + URI
-                // unchanged — no re-encode, no PUT.
-                out_list_entries.push(entry.clone());
-            }
-        }
-    }
-
-    // Cold partitions (touched but no prior entry): emit a
-    // fresh part with just the new superfiles.
-    for (pk, new_for_pk) in new_by_partition {
-        if handled_partitions.contains(&pk) {
-            continue;
-        }
-        let (fresh_entry, fresh_part, fresh_encoded) = build_part_and_entry(opts, new_for_pk, pk)?;
-        out_list_entries.push(fresh_entry);
-        parts_to_write.push(EncodedPart {
-            part: fresh_part,
-            encoded: fresh_encoded,
-        });
-    }
-
-    // At this point, out_list_entries contains all new ManifestListEntries that will be written.
-    // If these out_list_entries i.e Vec<ManifestListEntry> cause new ManifestParts to be created, those
-    // are stored in parts_to_write.
-
-    let mut out_list_entries_after_removal = Vec::new();
-    for entry in out_list_entries {
-        let Some(removals) = removals_by_partition.get(&entry.partition_key) else {
-            // If this entry belongs to a partition which has no removals, we can keep it as-is.
-            // This will also not need any change to parts_to_write.
-            out_list_entries_after_removal.push(entry);
-            continue;
-        };
-
-        let removal_ids = removals.iter().map(|r| r.superfile_id).collect::<Vec<_>>();
-        // TODO: Handle merging 2 parts into one if their sum is within threshold
-
-        // First we fetch the latest superfile entries - either from parts_to_write or the old manifest.
-        let (final_superfile_entries, existing_part_to_update) = if let Some(existing) =
-            parts_to_write
-                .iter_mut()
-                .find(|ep| ep.part.part_id == entry.part_id)
-        {
-            (
-                existing
-                    .part
-                    .superfiles
-                    .iter()
-                    .filter(|s| !removal_ids.contains(&s.superfile_id))
-                    .cloned()
-                    .collect::<Vec<_>>(),
-                Some(existing),
-            )
-        } else if let Ok(existing_part) = old.get_part_by_id(entry.part_id).await {
-            (
-                existing_part
-                    .superfiles
-                    .iter()
-                    .filter(|s| !removal_ids.contains(&s.superfile_id))
-                    .cloned()
-                    .collect::<Vec<_>>(),
-                None,
-            )
-        } else {
-            return Err(CommitError::PointerParse(
-                "Failed to find existing part for removal".to_string(),
-            ));
-        };
-
-        // Now we build the fresh part and entry based on the final superfile entries.
-        let (fresh_entry, fresh_part, fresh_encoded) =
-            build_part_and_entry(opts, final_superfile_entries, entry.partition_key)?;
-
-        if let Some(existing) = existing_part_to_update {
-            *existing = EncodedPart {
-                part: fresh_part,
-                encoded: fresh_encoded,
-            };
-        } else {
-            parts_to_write.push(EncodedPart {
-                part: fresh_part,
-                encoded: fresh_encoded,
-            });
-        }
-
-        out_list_entries_after_removal.push(fresh_entry);
-    }
-
-    Ok((out_list_entries_after_removal, parts_to_write))
 }
 
 #[cfg(test)]
@@ -700,11 +382,8 @@ mod tests {
     use super::*;
 
     use crate::storage::LocalFsStorageProvider;
-    use crate::supertable::{Manifest, SuperfileEntry, SuperfileUri};
-    use arrow_schema::{DataType, Field};
     use std::sync::Arc;
     use tempfile::TempDir;
-    use tokio::sync::OnceCell;
 
     // ---- URI helpers ---------------------------------------------------
 
@@ -1009,1744 +688,5 @@ mod tests {
         assert_eq!(POINTER_PATH, "_supertable/current");
         assert_eq!(MANIFEST_LISTS_DIR, "manifest-lists");
         assert_eq!(MANIFEST_PARTS_DIR, "manifests");
-    }
-
-    // ---- rebalance_for_commit ------------------------------------------
-
-    fn make_superfile_entry(docs: u64, pk: Vec<u8>) -> Arc<SuperfileEntry> {
-        Arc::new(SuperfileEntry {
-            superfile_id: uuid::Uuid::new_v4(),
-            uri: SuperfileUri::new_v4(),
-            n_docs: docs,
-            id_min: 0,
-            id_max: docs as i128 - 1,
-            scalar_stats: Default::default(),
-            fts_summary: Default::default(),
-            vector_summary: Default::default(),
-            partition_key: pk,
-            partition_hint: None,
-            subsection_offsets: None,
-        })
-    }
-
-    fn hash_bucket_0_pk() -> Vec<u8> {
-        // Hash partition with n_buckets=1 encodes to [0, 0, 0, 0] in little-endian
-        vec![0, 0, 0, 0]
-    }
-
-    fn simple_schema() -> std::sync::Arc<arrow_schema::Schema> {
-        std::sync::Arc::new(arrow_schema::Schema::new(vec![Field::new(
-            "text",
-            DataType::LargeUtf8,
-            false,
-        )]))
-    }
-
-    fn make_opts() -> std::sync::Arc<SupertableOptions> {
-        SupertableOptions::new(simple_schema(), vec![], vec![], None)
-            .map(Arc::new)
-            .expect("valid options")
-    }
-
-    fn empty_manifest(opts: &Arc<SupertableOptions>) -> Arc<Manifest> {
-        Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList::empty(opts.clone()),
-            list: Some(ManifestList {
-                format_version: list_mod::FORMAT_VERSION.into(),
-                manifest_id: 0,
-                options_hash: ContentHash([0u8; 32]),
-                schema: vec![],
-                id_column: "_id".into(),
-                fts_columns: vec![],
-                vector_columns: vec![],
-                partition_strategy: PartitionStrategy::Hash {
-                    column: "_id".into(),
-                    n_buckets: 1,
-                },
-                parts: vec![],
-            }),
-            parts: dashmap::DashMap::new(),
-            loader: None,
-        })
-    }
-
-    #[tokio::test]
-    async fn rebalance_fresh_start_cold_partition_should_create_entry() {
-        let opts = make_opts();
-        let old_manifest = empty_manifest(&opts);
-        let pk = hash_bucket_0_pk();
-
-        let new_entry = make_superfile_entry(100, pk.clone());
-        let new_entries = vec![new_entry];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        assert_eq!(parts[0].part.superfiles.len(), 1);
-        assert_eq!(parts[0].part.superfiles[0].n_docs, 100);
-    }
-
-    #[tokio::test]
-    async fn rebalance_fresh_start_multiple_cold_partitions_should_create_entries() {
-        // With Hash strategy (n_buckets=1), all entries map to the same partition.
-        let opts = make_opts();
-        let old_manifest = empty_manifest(&opts);
-        let pk = hash_bucket_0_pk();
-
-        let entry1 = make_superfile_entry(100, pk.clone());
-        let entry2 = make_superfile_entry(200, pk.clone());
-        let new_entries = vec![entry1, entry2];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 2);
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-        let total_docs: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 300);
-    }
-
-    #[tokio::test]
-    async fn rebalance_add_to_existing_partition_rewrites_part() {
-        // Adding a new entry to an existing single-part partition rewrites that part.
-        let opts = make_opts();
-        let pk_untouched = hash_bucket_0_pk();
-
-        let (_dir, storage) = local_storage();
-
-        let old_superfile = make_superfile_entry(100, pk_untouched.clone());
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![old_superfile.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 1,
-                partition_key: pk_untouched.clone(),
-                id_range: (0, 99),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-
-        let parts = dashmap::DashMap::new();
-        parts.insert(
-            pw.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![old_superfile],
-            },
-            list: Some(list),
-            parts,
-            loader: Some(Arc::new(loader)),
-        });
-
-        // Add new entry to the SAME partition (not a new/cold partition)
-        let new_entry = make_superfile_entry(50, pk_untouched.clone());
-        let new_entries = vec![new_entry];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // Should have 1 list entry (rewritten old one)
-        assert_eq!(list_entries.len(), 1);
-        // Should have 1 new part (the rewritten one)
-        assert_eq!(parts.len(), 1);
-
-        // Entry should be for the same partition
-        assert_eq!(list_entries[0].partition_key, pk_untouched);
-        assert_eq!(list_entries[0].n_superfiles, 2);
-
-        // Part should have combined superfiles
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-        let total_docs: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 150);
-    }
-
-    #[tokio::test]
-    async fn rebalance_rewrite_partition_within_target() {
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 3;
-        let opts = Arc::new(base_opts);
-
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf1 = make_superfile_entry(100, pk.clone());
-        let sf2 = make_superfile_entry(150, pk.clone());
-
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf1.clone(), sf2.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 2,
-                partition_key: pk.clone(),
-                id_range: (0, 149),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-
-        let parts = dashmap::DashMap::new();
-        parts.insert(
-            pw.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf1, sf2],
-            },
-            list: Some(list),
-            parts,
-            loader: Some(Arc::new(loader)),
-        });
-
-        // Add 1 new superfile to same partition (2 + 1 = 3, within target)
-        let new_entry = make_superfile_entry(75, pk.clone());
-        let new_entries = vec![new_entry];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // Rewrite case: 1 list entry (old entry replaced), 1 new part
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-
-        // Entry should be for same partition
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 3);
-
-        // Part should have all 3 superfiles combined
-        let part = &parts[0];
-        assert_eq!(part.part.superfiles.len(), 3);
-        // Verify combined doc count
-        let total_docs: u64 = part.part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 325); // 100 + 150 + 75
-    }
-
-    #[tokio::test]
-    async fn rebalance_split_partition_exceeds_target() {
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 2;
-        let opts = Arc::new(base_opts);
-
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf1 = make_superfile_entry(100, pk.clone());
-        let sf2 = make_superfile_entry(150, pk.clone());
-
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf1.clone(), sf2.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 2,
-                partition_key: pk.clone(),
-                id_range: (0, 149),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-
-        let parts = dashmap::DashMap::new();
-        parts.insert(
-            pw.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf1, sf2],
-            },
-            list: Some(list),
-            parts,
-            loader: Some(Arc::new(loader)),
-        });
-
-        // Add 2 new superfiles to same partition (2 + 2 = 4, exceeds target of 2)
-        let new_entry1 = make_superfile_entry(75, pk.clone());
-        let new_entry2 = make_superfile_entry(80, pk.clone());
-        let new_entries = vec![new_entry1, new_entry2];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // Split case: 2 list entries (old + fresh for split), 1 new part (fresh)
-        assert_eq!(list_entries.len(), 2);
-        assert_eq!(parts.len(), 1);
-
-        // Both entries should be for same partition
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[1].partition_key, pk);
-
-        // First entry (old) should still have original superfiles
-        assert_eq!(list_entries[0].n_superfiles, 2);
-
-        // Second entry (fresh) should have the new superfiles
-        assert_eq!(list_entries[1].n_superfiles, 2);
-
-        // The one new part should have exactly the 2 new superfiles
-        let part = &parts[0];
-        assert_eq!(part.part.superfiles.len(), 2);
-        let total_docs: u64 = part.part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 155); // 75 + 80
-    }
-
-    fn make_superfile_entry_hinted(docs: u64, pk: Vec<u8>, hint: u32) -> Arc<SuperfileEntry> {
-        Arc::new(SuperfileEntry {
-            superfile_id: uuid::Uuid::new_v4(),
-            uri: SuperfileUri::new_v4(),
-            n_docs: docs,
-            id_min: 0,
-            id_max: docs as i128 - 1,
-            scalar_stats: Default::default(),
-            fts_summary: Default::default(),
-            vector_summary: Default::default(),
-            partition_key: pk,
-            partition_hint: Some(hint),
-            subsection_offsets: None,
-        })
-    }
-
-    fn hash2_pk(bucket: u32) -> Vec<u8> {
-        bucket.to_le_bytes().to_vec()
-    }
-
-    #[tokio::test]
-    async fn rebalance_older_entry_preserved_when_latest_rewritten() {
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 2;
-        let opts = Arc::new(base_opts);
-
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf_old = make_superfile_entry(100, pk.clone());
-        let sf_latest = make_superfile_entry(150, pk.clone());
-
-        let part_old = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_old.clone()],
-        };
-        let pw_old = write_manifest_part(storage.as_ref(), &part_old, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_old");
-
-        let part_latest = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_latest.clone()],
-        };
-        let pw_latest = write_manifest_part(storage.as_ref(), &part_latest, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_latest");
-
-        // Old manifest with TWO entries for same partition (result of prior split)
-        // Second one is the "latest" for that partition
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_old.part_id,
-                    uri: pw_old.uri.clone(),
-                    content_hash: pw_old.content_hash,
-                    size_bytes_compressed: pw_old.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_old.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk.clone(),
-                    id_range: (0, 99),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_latest.part_id,
-                    uri: pw_latest.uri,
-                    content_hash: pw_latest.content_hash,
-                    size_bytes_compressed: pw_latest.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_latest.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk.clone(),
-                    id_range: (0, 149),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-
-        let parts = dashmap::DashMap::new();
-        parts.insert(
-            part_latest.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_latest)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_old, sf_latest],
-            },
-            list: Some(list),
-            parts,
-            loader: Some(Arc::new(loader)),
-        });
-
-        // Add one new entry for the partition
-        let new_entries = vec![make_superfile_entry(75, pk.clone())];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // Expect: old entry (preserved) + latest entry (rewritten) = 2 list entries
-        // Expect: 1 new part (latest rewrite)
-        assert_eq!(list_entries.len(), 2);
-        assert_eq!(parts.len(), 1);
-
-        // Both should be for same partition
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[1].partition_key, pk);
-
-        // First entry should carry over the old one unchanged
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        // URI should be exactly the same as the original written part
-        assert_eq!(list_entries[0].uri, pw_old.uri);
-
-        // Second entry should be the rewritten latest (1 + 1 = 2 superfiles)
-        assert_eq!(list_entries[1].n_superfiles, 2);
-
-        // New part should have the combined latest + new
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-        let total_docs: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 225); // 150 + 75
-    }
-
-    // ---- cross-partition tests --------------------------------------------
-
-    #[tokio::test]
-    async fn rebalance_two_partitions_both_touched() {
-        // Two distinct partitions each have one existing superfile; a new
-        // entry is added to both. Both should be rewritten independently.
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 3;
-        let opts = Arc::new(base_opts);
-
-        let pk_a = hash2_pk(0);
-        let pk_b = hash2_pk(1);
-        let (_dir, storage) = local_storage();
-
-        let sf_a = make_superfile_entry_hinted(100, pk_a.clone(), 0);
-        let part_a = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a.clone()],
-        };
-        let pw_a = write_manifest_part(storage.as_ref(), &part_a, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_a");
-
-        let sf_b = make_superfile_entry_hinted(200, pk_b.clone(), 1);
-        let part_b = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_b.clone()],
-        };
-        let pw_b = write_manifest_part(storage.as_ref(), &part_b, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_b");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 2,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_a.part_id,
-                    uri: pw_a.uri,
-                    content_hash: pw_a.content_hash,
-                    size_bytes_compressed: pw_a.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_a.clone(),
-                    id_range: (0, 99),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_b.part_id,
-                    uri: pw_b.uri,
-                    content_hash: pw_b.content_hash,
-                    size_bytes_compressed: pw_b.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_b.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_b.clone(),
-                    id_range: (0, 199),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            part_a.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a)))),
-        );
-        parts_map.insert(
-            part_b.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_b)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_a, sf_b],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let new_entries = vec![
-            make_superfile_entry_hinted(50, pk_a.clone(), 0),
-            make_superfile_entry_hinted(80, pk_b.clone(), 1),
-        ];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // Both partitions are rewritten: 2 list entries, 2 new parts
-        assert_eq!(list_entries.len(), 2);
-        assert_eq!(parts.len(), 2);
-
-        // Order preserved: partition A first, then B
-        assert_eq!(list_entries[0].partition_key, pk_a);
-        assert_eq!(list_entries[1].partition_key, pk_b);
-
-        // Partition A: 1 existing + 1 new = 2 superfiles, 150 docs
-        assert_eq!(list_entries[0].n_superfiles, 2);
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-        let docs_a: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(docs_a, 150);
-
-        // Partition B: 1 existing + 1 new = 2 superfiles, 280 docs
-        assert_eq!(list_entries[1].n_superfiles, 2);
-        assert_eq!(parts[1].part.superfiles.len(), 2);
-        let docs_b: u64 = parts[1].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(docs_b, 280);
-    }
-
-    #[tokio::test]
-    async fn rebalance_two_partitions_one_touched_exact_carry_over() {
-        // Partition A is touched (gets a new entry); partition B is not.
-        // Verifies that B's list entry carries over with the exact URI and
-        // content_hash that were written — no re-encode, no PUT.
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 3;
-        let opts = Arc::new(base_opts);
-
-        let pk_a = hash2_pk(0);
-        let pk_b = hash2_pk(1);
-        let (_dir, storage) = local_storage();
-
-        let sf_a = make_superfile_entry_hinted(100, pk_a.clone(), 0);
-        let part_a = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a.clone()],
-        };
-        let pw_a = write_manifest_part(storage.as_ref(), &part_a, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_a");
-
-        let sf_b = make_superfile_entry_hinted(200, pk_b.clone(), 1);
-        let part_b = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_b.clone()],
-        };
-        let pw_b = write_manifest_part(storage.as_ref(), &part_b, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_b");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 2,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_a.part_id,
-                    uri: pw_a.uri,
-                    content_hash: pw_a.content_hash,
-                    size_bytes_compressed: pw_a.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_a.clone(),
-                    id_range: (0, 99),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_b.part_id,
-                    uri: pw_b.uri.clone(),
-                    content_hash: pw_b.content_hash,
-                    size_bytes_compressed: pw_b.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_b.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_b.clone(),
-                    id_range: (0, 199),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            part_a.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a)))),
-        );
-        parts_map.insert(
-            part_b.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_b)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_a, sf_b],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        // Only touch partition A
-        let new_entries = vec![make_superfile_entry_hinted(50, pk_a.clone(), 0)];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // 2 list entries (A rewritten, B carried over), 1 new part (A only)
-        assert_eq!(list_entries.len(), 2);
-        assert_eq!(parts.len(), 1);
-
-        // Partition A: rewritten with 2 superfiles, 150 docs
-        assert_eq!(list_entries[0].partition_key, pk_a);
-        assert_eq!(list_entries[0].n_superfiles, 2);
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-        let docs_a: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(docs_a, 150);
-
-        // Partition B: exact carry-over — URI and content_hash unchanged
-        assert_eq!(list_entries[1].partition_key, pk_b);
-        assert_eq!(list_entries[1].n_superfiles, 1);
-        assert_eq!(list_entries[1].uri, pw_b.uri);
-        assert_eq!(list_entries[1].content_hash, pw_b.content_hash);
-    }
-
-    #[tokio::test]
-    async fn rebalance_two_partitions_each_with_prior_split() {
-        // Each partition already has two parts from a prior split: an older
-        // frozen part and a latest mutable part. Adding one new entry to each
-        // partition should rewrite only the latest part for each, carrying
-        // the older parts over unchanged.
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 2;
-        let opts = Arc::new(base_opts);
-
-        let pk_a = hash2_pk(0);
-        let pk_b = hash2_pk(1);
-        let (_dir, storage) = local_storage();
-
-        // Partition A: two parts
-        let sf_a_old = make_superfile_entry_hinted(100, pk_a.clone(), 0);
-        let part_a_old = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_old.clone()],
-        };
-        let pw_a_old = write_manifest_part(storage.as_ref(), &part_a_old, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_a_old");
-
-        let sf_a_latest = make_superfile_entry_hinted(150, pk_a.clone(), 0);
-        let part_a_latest = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_latest.clone()],
-        };
-        let pw_a_latest =
-            write_manifest_part(storage.as_ref(), &part_a_latest, MANIFEST_ZSTD_LEVEL)
-                .await
-                .expect("write part_a_latest");
-
-        // Partition B: two parts
-        let sf_b_old = make_superfile_entry_hinted(200, pk_b.clone(), 1);
-        let part_b_old = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_b_old.clone()],
-        };
-        let pw_b_old = write_manifest_part(storage.as_ref(), &part_b_old, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_b_old");
-
-        let sf_b_latest = make_superfile_entry_hinted(250, pk_b.clone(), 1);
-        let part_b_latest = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_b_latest.clone()],
-        };
-        let pw_b_latest =
-            write_manifest_part(storage.as_ref(), &part_b_latest, MANIFEST_ZSTD_LEVEL)
-                .await
-                .expect("write part_b_latest");
-
-        // List order: [a_old, a_latest, b_old, b_latest]
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 2,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_a_old.part_id,
-                    uri: pw_a_old.uri.clone(),
-                    content_hash: pw_a_old.content_hash,
-                    size_bytes_compressed: pw_a_old.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a_old.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_a.clone(),
-                    id_range: (0, 99),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_a_latest.part_id,
-                    uri: pw_a_latest.uri,
-                    content_hash: pw_a_latest.content_hash,
-                    size_bytes_compressed: pw_a_latest.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a_latest.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_a.clone(),
-                    id_range: (0, 149),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_b_old.part_id,
-                    uri: pw_b_old.uri.clone(),
-                    content_hash: pw_b_old.content_hash,
-                    size_bytes_compressed: pw_b_old.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_b_old.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_b.clone(),
-                    id_range: (0, 199),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_b_latest.part_id,
-                    uri: pw_b_latest.uri,
-                    content_hash: pw_b_latest.content_hash,
-                    size_bytes_compressed: pw_b_latest.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_b_latest.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_b.clone(),
-                    id_range: (0, 249),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            part_a_latest.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a_latest)))),
-        );
-        parts_map.insert(
-            part_b_latest.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_b_latest)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_a_old, sf_a_latest, sf_b_old, sf_b_latest],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let new_entries = vec![
-            make_superfile_entry_hinted(75, pk_a.clone(), 0),
-            make_superfile_entry_hinted(90, pk_b.clone(), 1),
-        ];
-
-        let (list_entries, parts) = rebalance_for_commit(&opts, &old_manifest, &new_entries, &[])
-            .await
-            .expect("rebalance");
-
-        // 4 list entries: [a_old, a_rewritten, b_old, b_rewritten]
-        assert_eq!(list_entries.len(), 4);
-        // 2 new parts: one rewrite per partition
-        assert_eq!(parts.len(), 2);
-
-        // [0] Partition A old: carried over exactly — URI and content_hash unchanged
-        assert_eq!(list_entries[0].partition_key, pk_a);
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        assert_eq!(list_entries[0].uri, pw_a_old.uri);
-        assert_eq!(list_entries[0].content_hash, pw_a_old.content_hash);
-
-        // [1] Partition A latest: rewritten with 1 existing + 1 new = 2 superfiles, 225 docs
-        assert_eq!(list_entries[1].partition_key, pk_a);
-        assert_eq!(list_entries[1].n_superfiles, 2);
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-        let docs_a: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(docs_a, 225); // 150 + 75
-
-        // [2] Partition B old: carried over exactly — URI and content_hash unchanged
-        assert_eq!(list_entries[2].partition_key, pk_b);
-        assert_eq!(list_entries[2].n_superfiles, 1);
-        assert_eq!(list_entries[2].uri, pw_b_old.uri);
-        assert_eq!(list_entries[2].content_hash, pw_b_old.content_hash);
-
-        // [3] Partition B latest: rewritten with 1 existing + 1 new = 2 superfiles, 340 docs
-        assert_eq!(list_entries[3].partition_key, pk_b);
-        assert_eq!(list_entries[3].n_superfiles, 2);
-        assert_eq!(parts[1].part.superfiles.len(), 2);
-        let docs_b: u64 = parts[1].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(docs_b, 340); // 250 + 90
-    }
-
-    // ---- removal tests ---------------------------------------------------
-
-    #[tokio::test]
-    async fn rebalance_remove_one_superfile_from_partition() {
-        // Partition has 2 superfiles; remove one. Verifies the part is
-        // rewritten containing only the superfile that was not removed.
-        let opts = make_opts();
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf_keep = make_superfile_entry(100, pk.clone());
-        let sf_remove = make_superfile_entry(150, pk.clone());
-
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_keep.clone(), sf_remove.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 2,
-                partition_key: pk.clone(),
-                id_range: (0, 149),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            existing_part.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_keep.clone(), sf_remove.clone()],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let (list_entries, parts) =
-            rebalance_for_commit(&opts, &old_manifest, &[], std::slice::from_ref(&sf_remove))
-                .await
-                .expect("rebalance");
-
-        // Part rewritten with 1 superfile; no cold entries
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        assert_eq!(parts[0].part.superfiles.len(), 1);
-        assert_eq!(
-            parts[0].part.superfiles[0].superfile_id,
-            sf_keep.superfile_id
-        );
-        let total_docs: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 100);
-    }
-
-    #[tokio::test]
-    async fn rebalance_add_and_remove_in_same_partition() {
-        // One new superfile is added while one existing superfile is removed
-        // in the same partition. The resulting part should contain the
-        // surviving existing superfile plus the new one — not the removed one.
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 3;
-        let opts = Arc::new(base_opts);
-
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf_keep = make_superfile_entry(100, pk.clone());
-        let sf_remove = make_superfile_entry(150, pk.clone());
-
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_keep.clone(), sf_remove.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 2,
-                partition_key: pk.clone(),
-                id_range: (0, 149),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            existing_part.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_keep.clone(), sf_remove.clone()],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let sf_new = make_superfile_entry(75, pk.clone());
-        let new_entries = vec![sf_new.clone()];
-
-        let (list_entries, parts) = rebalance_for_commit(
-            &opts,
-            &old_manifest,
-            &new_entries,
-            std::slice::from_ref(&sf_remove),
-        )
-        .await
-        .expect("rebalance");
-
-        // Net result: 1 list entry, 1 part — sf_keep + sf_new, sf_remove absent
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 2);
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-
-        let ids: Vec<_> = parts[0]
-            .part
-            .superfiles
-            .iter()
-            .map(|s| s.superfile_id)
-            .collect();
-        assert!(ids.contains(&sf_keep.superfile_id));
-        assert!(ids.contains(&sf_new.superfile_id));
-        assert!(!ids.contains(&sf_remove.superfile_id));
-
-        let total_docs: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(total_docs, 175); // 100 + 75
-    }
-
-    #[tokio::test]
-    async fn rebalance_remove_from_one_partition_other_carried_over_exactly() {
-        // Two partitions: remove a superfile from partition A, leave partition B alone.
-        // Verifies partition B's list entry is carried over with the exact URI and
-        // content_hash — no re-encode, no PUT — while partition A is rewritten.
-        let opts = make_opts();
-        let pk_a = hash2_pk(0);
-        let pk_b = hash2_pk(1);
-        let (_dir, storage) = local_storage();
-
-        let sf_a_keep = make_superfile_entry_hinted(100, pk_a.clone(), 0);
-        let sf_a_remove = make_superfile_entry_hinted(150, pk_a.clone(), 0);
-        let part_a = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_keep.clone(), sf_a_remove.clone()],
-        };
-        let pw_a = write_manifest_part(storage.as_ref(), &part_a, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_a");
-
-        let sf_b = make_superfile_entry_hinted(200, pk_b.clone(), 1);
-        let part_b = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_b.clone()],
-        };
-        let pw_b = write_manifest_part(storage.as_ref(), &part_b, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_b");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 2,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_a.part_id,
-                    uri: pw_a.uri,
-                    content_hash: pw_a.content_hash,
-                    size_bytes_compressed: pw_a.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a.size_bytes_uncompressed,
-                    n_superfiles: 2,
-                    partition_key: pk_a.clone(),
-                    id_range: (0, 149),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_b.part_id,
-                    uri: pw_b.uri.clone(),
-                    content_hash: pw_b.content_hash,
-                    size_bytes_compressed: pw_b.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_b.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk_b.clone(),
-                    id_range: (0, 199),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            part_a.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a)))),
-        );
-        parts_map.insert(
-            part_b.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_b)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf_a_keep.clone(), sf_a_remove.clone(), sf_b.clone()],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let (list_entries, parts) = rebalance_for_commit(
-            &opts,
-            &old_manifest,
-            &[],
-            std::slice::from_ref(&sf_a_remove),
-        )
-        .await
-        .expect("rebalance");
-
-        // 2 list entries, 1 new part (only partition A was rewritten)
-        assert_eq!(list_entries.len(), 2);
-        assert_eq!(parts.len(), 1);
-
-        // Partition A: rewritten with 1 surviving superfile
-        assert_eq!(list_entries[0].partition_key, pk_a);
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        assert_eq!(parts[0].part.superfiles.len(), 1);
-        assert_eq!(
-            parts[0].part.superfiles[0].superfile_id,
-            sf_a_keep.superfile_id
-        );
-        let docs_a: u64 = parts[0].part.superfiles.iter().map(|s| s.n_docs).sum();
-        assert_eq!(docs_a, 100);
-
-        // Partition B: exact carry-over — URI and content_hash unchanged
-        assert_eq!(list_entries[1].partition_key, pk_b);
-        assert_eq!(list_entries[1].n_superfiles, 1);
-        assert_eq!(list_entries[1].uri, pw_b.uri);
-        assert_eq!(list_entries[1].content_hash, pw_b.content_hash);
-    }
-
-    #[tokio::test]
-    async fn rebalance_remove_from_latest_part_in_split_partition() {
-        // Partition A has two parts from a prior split: part_a_old (frozen, 1 sf)
-        // and part_a_latest (mutable, 2 sfs). We remove sf_a_latest_remove,
-        // which lives in the SECOND (latest) part.
-        //
-        // Bug: the removal loop calls removals_by_partition.remove(&partition_key)
-        // for each entry in out_list_entries. When part_a_old is processed first,
-        // the key [0,0,0,0] is consumed from the map. When part_a_latest is
-        // processed second, remove() returns None and the entry carries over
-        // unchanged — sf_a_latest_remove is never removed. As a side effect,
-        // part_a_old is unnecessarily rewritten (its URI changes even though its
-        // contents did not).
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 2;
-        let opts = Arc::new(base_opts);
-
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        // part_a_old: frozen entry from a prior split
-        let sf_a_old = make_superfile_entry(100, pk.clone());
-        let part_a_old = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_old.clone()],
-        };
-        let pw_a_old = write_manifest_part(storage.as_ref(), &part_a_old, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_a_old");
-
-        // part_a_latest: current mutable entry; contains the sf to remove
-        let sf_a_latest_keep = make_superfile_entry(150, pk.clone());
-        let sf_a_latest_remove = make_superfile_entry(200, pk.clone());
-        let part_a_latest = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_latest_keep.clone(), sf_a_latest_remove.clone()],
-        };
-        let pw_a_latest =
-            write_manifest_part(storage.as_ref(), &part_a_latest, MANIFEST_ZSTD_LEVEL)
-                .await
-                .expect("write part_a_latest");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_a_old.part_id,
-                    uri: pw_a_old.uri.clone(),
-                    content_hash: pw_a_old.content_hash,
-                    size_bytes_compressed: pw_a_old.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a_old.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk.clone(),
-                    id_range: (0, 99),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_a_latest.part_id,
-                    uri: pw_a_latest.uri.clone(),
-                    content_hash: pw_a_latest.content_hash,
-                    size_bytes_compressed: pw_a_latest.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a_latest.size_bytes_uncompressed,
-                    n_superfiles: 2,
-                    partition_key: pk.clone(),
-                    id_range: (0, 199),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            part_a_old.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a_old)))),
-        );
-        parts_map.insert(
-            part_a_latest.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a_latest)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![
-                    sf_a_old.clone(),
-                    sf_a_latest_keep.clone(),
-                    sf_a_latest_remove.clone(),
-                ],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let (list_entries, parts) = rebalance_for_commit(
-            &opts,
-            &old_manifest,
-            &[],
-            std::slice::from_ref(&sf_a_latest_remove),
-        )
-        .await
-        .expect("rebalance");
-
-        assert_eq!(list_entries.len(), 2);
-        // Both parts in the split are rewritten: any part in a partition with a
-        // pending removal is rewritten regardless of whether the removal matched
-        // anything in it.
-        assert_eq!(parts.len(), 2);
-
-        // Both list entries are for the same partition
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[1].partition_key, pk);
-
-        // sf_a_old survives (in one of the output parts)
-        // sf_a_latest_keep survives (in one of the output parts)
-        // sf_a_latest_remove is absent from every output part
-        let all_ids: Vec<_> = parts
-            .iter()
-            .flat_map(|ep| ep.part.superfiles.iter())
-            .map(|s| s.superfile_id)
-            .collect();
-        assert!(
-            all_ids.contains(&sf_a_old.superfile_id),
-            "sf_a_old must survive"
-        );
-        assert!(
-            all_ids.contains(&sf_a_latest_keep.superfile_id),
-            "sf_a_latest_keep must survive"
-        );
-        assert!(
-            !all_ids.contains(&sf_a_latest_remove.superfile_id),
-            "sf_a_latest_remove must be absent"
-        );
-
-        // Each rewritten part has exactly 1 superfile
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        assert_eq!(list_entries[1].n_superfiles, 1);
-    }
-
-    #[tokio::test]
-    async fn rebalance_remove_all_superfiles_empties_partition() {
-        // All superfiles in a partition are removed. Documents the current
-        // behavior: the list entry survives with n_superfiles=0 and the
-        // part has no superfiles (empty partition).
-        let opts = make_opts();
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf1 = make_superfile_entry(100, pk.clone());
-        let sf2 = make_superfile_entry(150, pk.clone());
-
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf1.clone(), sf2.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 2,
-                partition_key: pk.clone(),
-                id_range: (0, 149),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            existing_part.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf1.clone(), sf2.clone()],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let (list_entries, parts) =
-            rebalance_for_commit(&opts, &old_manifest, &[], &[sf1.clone(), sf2.clone()])
-                .await
-                .expect("rebalance");
-
-        // Both superfiles removed: list entry remains with n_superfiles=0
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 0);
-        assert_eq!(parts[0].part.superfiles.len(), 0);
-    }
-
-    #[tokio::test]
-    async fn rebalance_remove_nonexistent_superfile_id_is_noop() {
-        // entries_to_remove contains a superfile_id that is not present in any
-        // part. The filter matches nothing and both original superfiles survive.
-        // The part is still rewritten (the removal loop doesn't skip parts where
-        // no removal matched), so n_superfiles stays at 2.
-        let opts = make_opts();
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        let sf1 = make_superfile_entry(100, pk.clone());
-        let sf2 = make_superfile_entry(150, pk.clone());
-
-        let existing_part = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf1.clone(), sf2.clone()],
-        };
-        let pw = write_manifest_part(storage.as_ref(), &existing_part, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![ManifestListEntry {
-                part_id: pw.part_id,
-                uri: pw.uri,
-                content_hash: pw.content_hash,
-                size_bytes_compressed: pw.size_bytes_compressed,
-                size_bytes_uncompressed: pw.size_bytes_uncompressed,
-                n_superfiles: 2,
-                partition_key: pk.clone(),
-                id_range: (0, 149),
-                scalar_stats_agg: Default::default(),
-                fts_summary_agg: Default::default(),
-                vector_summary_agg: Default::default(),
-            }],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            existing_part.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(existing_part)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![sf1.clone(), sf2.clone()],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        // sf_ghost was never added to any part; its superfile_id won't match anything
-        let sf_ghost = make_superfile_entry(50, pk.clone());
-
-        let (list_entries, parts) =
-            rebalance_for_commit(&opts, &old_manifest, &[], std::slice::from_ref(&sf_ghost))
-                .await
-                .expect("rebalance");
-
-        assert_eq!(list_entries.len(), 1);
-        assert_eq!(parts.len(), 1);
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[0].n_superfiles, 2);
-        assert_eq!(parts[0].part.superfiles.len(), 2);
-
-        let ids: Vec<_> = parts[0]
-            .part
-            .superfiles
-            .iter()
-            .map(|s| s.superfile_id)
-            .collect();
-        assert!(ids.contains(&sf1.superfile_id), "sf1 must survive");
-        assert!(ids.contains(&sf2.superfile_id), "sf2 must survive");
-        assert!(
-            !ids.contains(&sf_ghost.superfile_id),
-            "ghost id must not appear"
-        );
-    }
-
-    #[tokio::test]
-    async fn rebalance_remove_from_older_frozen_part_in_split_partition() {
-        // Partition A has two parts from a prior split: part_a_old (frozen, 2
-        // sfs: sf_a_old_keep + sf_a_old_remove) and part_a_latest (mutable, 1
-        // sf). We remove sf_a_old_remove, which lives in the FIRST (older,
-        // frozen) part.
-        //
-        // Because the fix applies the removal set to every part in the partition,
-        // both parts are rewritten. sf_a_old_remove is absent from the output;
-        // sf_a_old_keep and sf_a_latest survive.
-        let mut base_opts =
-            SupertableOptions::new(simple_schema(), vec![], vec![], None).expect("valid options");
-        base_opts.target_superfiles_per_partition = 2;
-        let opts = Arc::new(base_opts);
-
-        let pk = hash_bucket_0_pk();
-        let (_dir, storage) = local_storage();
-
-        // part_a_old: frozen entry — contains the sf to remove
-        let sf_a_old_keep = make_superfile_entry(100, pk.clone());
-        let sf_a_old_remove = make_superfile_entry(150, pk.clone());
-        let part_a_old = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_old_keep.clone(), sf_a_old_remove.clone()],
-        };
-        let pw_a_old = write_manifest_part(storage.as_ref(), &part_a_old, MANIFEST_ZSTD_LEVEL)
-            .await
-            .expect("write part_a_old");
-
-        // part_a_latest: mutable entry — does not contain the sf to remove
-        let sf_a_latest = make_superfile_entry(200, pk.clone());
-        let part_a_latest = ManifestPart {
-            format_version: part_mod::FORMAT_VERSION.into(),
-            part_id: PartId::new_v4(),
-            superfiles: vec![sf_a_latest.clone()],
-        };
-        let pw_a_latest =
-            write_manifest_part(storage.as_ref(), &part_a_latest, MANIFEST_ZSTD_LEVEL)
-                .await
-                .expect("write part_a_latest");
-
-        let list = ManifestList {
-            format_version: list_mod::FORMAT_VERSION.into(),
-            manifest_id: 0,
-            options_hash: ContentHash([0u8; 32]),
-            schema: vec![],
-            id_column: "_id".into(),
-            fts_columns: vec![],
-            vector_columns: vec![],
-            partition_strategy: PartitionStrategy::Hash {
-                column: "_id".into(),
-                n_buckets: 1,
-            },
-            parts: vec![
-                ManifestListEntry {
-                    part_id: pw_a_old.part_id,
-                    uri: pw_a_old.uri,
-                    content_hash: pw_a_old.content_hash,
-                    size_bytes_compressed: pw_a_old.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a_old.size_bytes_uncompressed,
-                    n_superfiles: 2,
-                    partition_key: pk.clone(),
-                    id_range: (0, 149),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-                ManifestListEntry {
-                    part_id: pw_a_latest.part_id,
-                    uri: pw_a_latest.uri,
-                    content_hash: pw_a_latest.content_hash,
-                    size_bytes_compressed: pw_a_latest.size_bytes_compressed,
-                    size_bytes_uncompressed: pw_a_latest.size_bytes_uncompressed,
-                    n_superfiles: 1,
-                    partition_key: pk.clone(),
-                    id_range: (0, 199),
-                    scalar_stats_agg: Default::default(),
-                    fts_summary_agg: Default::default(),
-                    vector_summary_agg: Default::default(),
-                },
-            ],
-        };
-        let loader = crate::supertable::manifest::ManifestPartLoader::new(storage, &list);
-        let parts_map = dashmap::DashMap::new();
-        parts_map.insert(
-            part_a_old.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a_old)))),
-        );
-        parts_map.insert(
-            part_a_latest.part_id,
-            Arc::new(OnceCell::new_with(Some(Arc::new(part_a_latest)))),
-        );
-        let old_manifest = Arc::new(Manifest {
-            superfile_list: crate::supertable::manifest::SuperfileList {
-                manifest_id: 0,
-                options: opts.clone(),
-                superfiles: vec![
-                    sf_a_old_keep.clone(),
-                    sf_a_old_remove.clone(),
-                    sf_a_latest.clone(),
-                ],
-            },
-            list: Some(list),
-            parts: parts_map,
-            loader: Some(Arc::new(loader)),
-        });
-
-        let (list_entries, parts) = rebalance_for_commit(
-            &opts,
-            &old_manifest,
-            &[],
-            std::slice::from_ref(&sf_a_old_remove),
-        )
-        .await
-        .expect("rebalance");
-
-        assert_eq!(list_entries.len(), 2);
-        // Both parts rewritten: the fix applies the removal set to every part in
-        // the partition, so the latest is also rewritten (no match, same content)
-        assert_eq!(parts.len(), 2);
-
-        assert_eq!(list_entries[0].partition_key, pk);
-        assert_eq!(list_entries[1].partition_key, pk);
-
-        // sf_a_old_keep and sf_a_latest survive; sf_a_old_remove is absent
-        let all_ids: Vec<_> = parts
-            .iter()
-            .flat_map(|ep| ep.part.superfiles.iter())
-            .map(|s| s.superfile_id)
-            .collect();
-        assert!(
-            all_ids.contains(&sf_a_old_keep.superfile_id),
-            "sf_a_old_keep must survive"
-        );
-        assert!(
-            all_ids.contains(&sf_a_latest.superfile_id),
-            "sf_a_latest must survive"
-        );
-        assert!(
-            !all_ids.contains(&sf_a_old_remove.superfile_id),
-            "sf_a_old_remove must be absent"
-        );
-
-        // Old part now has 1 sf (sf_a_old_remove was removed)
-        assert_eq!(list_entries[0].n_superfiles, 1);
-        // Latest part still has 1 sf (removal did not touch it)
-        assert_eq!(list_entries[1].n_superfiles, 1);
     }
 }
