+++ src/supertable/query/skip.rs
@@ -452,24 +452,13 @@ mod tests {
         Arc::new(e)
     }
 
-    fn manifest_with(
-        opts: Arc<SupertableOptions>,
-        superfiles: Vec<Arc<SuperfileEntry>>,
-    ) -> Manifest {
-        // build via `with_appended` so the new outer
-        // Manifest's metadata fields (list, parts, loader) get
-        // initialized correctly. Equivalent to the old direct-
-        // field-assignment helper for the test's purposes.
-        Manifest::empty(opts).with_appended(superfiles)
-    }
-
     // ---- fts_bloom_skip ----------------------------------------------
 
     #[test]
     fn bloom_skip_keeps_superfiles_with_any_query_term_in_or_mode() {
         let s_a = superfile_with_terms("title", &["alpha", "beta"]);
         let s_b = superfile_with_terms("title", &["gamma", "delta"]);
-        let m = manifest_with(opts_simple(), vec![s_a, s_b]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s_a, s_b]);
         let mask = fts_bloom_skip(&m.superfiles, "title", &["alpha", "missing"], BoolMode::Or);
         // Superfile A has alpha → keep. Superfile B has neither → prune.
         assert_eq!(mask, vec![true, false]);
@@ -479,7 +468,7 @@ mod tests {
     fn bloom_skip_requires_all_terms_present_in_and_mode() {
         let s_a = superfile_with_terms("title", &["alpha", "beta"]);
         let s_b = superfile_with_terms("title", &["alpha", "gamma"]);
-        let m = manifest_with(opts_simple(), vec![s_a, s_b]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s_a, s_b]);
         let mask = fts_bloom_skip(&m.superfiles, "title", &["alpha", "beta"], BoolMode::And);
         // Superfile A has both. Superfile B is missing 'beta' → prune.
         assert_eq!(mask, vec![true, false]);
@@ -488,7 +477,7 @@ mod tests {
     #[test]
     fn bloom_skip_unknown_column_keeps_all() {
         let s = superfile_with_terms("title", &["alpha"]);
-        let m = manifest_with(opts_simple(), vec![s]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s]);
         let mask = fts_bloom_skip(&m.superfiles, "no_such_column", &["alpha"], BoolMode::Or);
         assert_eq!(mask, vec![true]);
     }
@@ -496,14 +485,14 @@ mod tests {
     #[test]
     fn bloom_skip_empty_terms_keeps_all() {
         let s = superfile_with_terms("title", &["alpha"]);
-        let m = manifest_with(opts_simple(), vec![s]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s]);
         let mask = fts_bloom_skip(&m.superfiles, "title", &[], BoolMode::Or);
         assert_eq!(mask, vec![true]);
     }
 
     #[test]
     fn bloom_skip_with_no_superfiles_returns_empty_vec() {
-        let m = manifest_with(opts_simple(), vec![]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![]);
         let mask = fts_bloom_skip(&m.superfiles, "title", &["alpha"], BoolMode::Or);
         assert!(mask.is_empty());
     }
@@ -518,7 +507,7 @@ mod tests {
         //            overlaps the upper end.
         let s_a = superfile_with_terms("title", &["apple", "banana"]);
         let s_b = superfile_with_terms("title", &["python", "rust"]);
-        let m = manifest_with(opts_simple(), vec![s_a, s_b]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s_a, s_b]);
         let mask = fts_prefix_skip(&m.superfiles, "title", b"rust");
         assert_eq!(mask, vec![false, true]);
     }
@@ -527,7 +516,7 @@ mod tests {
     fn prefix_skip_keeps_superfiles_with_matching_prefix_inside_range() {
         // Terms ['rusting', 'rusty'] → prefix "rust" overlaps.
         let s = superfile_with_terms("title", &["rusting", "rusty"]);
-        let m = manifest_with(opts_simple(), vec![s]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s]);
         let mask = fts_prefix_skip(&m.superfiles, "title", b"rust");
         assert_eq!(mask, vec![true]);
     }
@@ -535,7 +524,7 @@ mod tests {
     #[test]
     fn prefix_skip_empty_prefix_keeps_all() {
         let s = superfile_with_terms("title", &["alpha"]);
-        let m = manifest_with(opts_simple(), vec![s]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s]);
         let mask = fts_prefix_skip(&m.superfiles, "title", b"");
         assert_eq!(mask, vec![true]);
     }
@@ -543,7 +532,7 @@ mod tests {
     #[test]
     fn prefix_skip_unknown_column_keeps_all() {
         let s = superfile_with_terms("title", &["alpha"]);
-        let m = manifest_with(opts_simple(), vec![s]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s]);
         let mask = fts_prefix_skip(&m.superfiles, "no_such_column", b"alp");
         assert_eq!(mask, vec![true]);
     }
@@ -552,7 +541,7 @@ mod tests {
     fn prefix_skip_zero_term_superfile_pruned() {
         // Empty term_range = no terms indexed. Prefix can't match.
         let s = Arc::new(empty_superfile());
-        let m = manifest_with(opts_simple(), vec![s]);
+        let m = Manifest::new_from_superfiles(opts_simple(), vec![s]);
         let mask = fts_prefix_skip(&m.superfiles, "title", b"rust");
         // No FTS summary on the superfile → keep (column-missing
         // path). Sanity: this is the "unknown column" path, not
@@ -566,7 +555,7 @@ mod tests {
     fn vector_centroid_skip_v1_keeps_all_superfiles() {
         let s_a = superfile_with_centroid("emb", vec![0.0; 16], 0.5);
         let s_b = superfile_with_centroid("emb", vec![10.0; 16], 0.5);
-        let m = manifest_with(opts_with_vector(), vec![s_a, s_b]);
+        let m = Manifest::new_from_superfiles(opts_with_vector(), vec![s_a, s_b]);
         let q = vec![0.0f32; 16];
         let mask = vector_centroid_skip(&m, "emb", &q);
         assert_eq!(mask, vec![true, true]);
@@ -594,7 +583,7 @@ mod tests {
             },
             0.0,
         );
-        let m = manifest_with(opts, vec![far.clone(), near.clone()]);
+        let m = Manifest::new_from_superfiles(opts, vec![far.clone(), near.clone()]);
         let q = {
             let mut v = vec![0.0f32; 16];
             v[0] = 1.0;
@@ -609,7 +598,7 @@ mod tests {
     fn superfiles_sorted_by_centroid_distance_pushes_missing_summary_to_end() {
         let with_v = superfile_with_centroid("emb", vec![1.0f32; 16], 0.0);
         let without_v = Arc::new(empty_superfile());
-        let m = manifest_with(opts_with_vector(), vec![without_v, with_v]);
+        let m = Manifest::new_from_superfiles(opts_with_vector(), vec![without_v, with_v]);
         let q = vec![1.0f32; 16];
         let order = superfiles_sorted_by_centroid_distance(&m, "emb", &q, Metric::L2Sq);
         // Index 1 (has summary) sorted before index 0 (missing).
