+++ tests/supertable/commit/stats.rs
@@ -6,16 +6,14 @@
 //! Covers:
 //!   - Fresh `create` returns the empty-supertable stats:
 //!     `manifest_id == 0`, `n_superfiles == 0`,
-//!     `n_manifest_parts == None`.
+//!     `n_manifest_parts == 0`.
 //!   - Stats track commits: each `writer.commit()` advances
 //!     `manifest_id` + grows `n_superfiles`.
 //!   - With storage attached, after the first commit
-//!     `n_manifest_parts == Some(1)` and
-//!     `n_manifest_parts_loaded == 0` (the writer's commit
-//!     path doesn't currently hydrate the part it just wrote
-//!     — `Supertable::open` is what populates the parts
-//!     cache; the writer just rebuilds the in-memory state
-//!     from `new_superfile_list`).
+//!     `n_manifest_parts == 1` and `n_manifest_parts_loaded == 1`:
+//!     the commit path seeds the freshly-written part(s) into the
+//!     in-memory cache, so a same-handle query serves them without
+//!     re-reading any manifest part from storage.
 //!   - `Supertable::open`'s eager-fetch populates
 //!     `n_manifest_parts_loaded == n_manifest_parts`.
 //!   - `process_rss_bytes` is non-zero and falls within a
@@ -101,10 +99,10 @@ fn stats_show_manifest_parts_when_storage_attached() {
         w.commit().expect("commit");
     }
 
-    // Producer's in-memory state after commit: list is set,
-    // parts cache is empty (writer rebuilds state via
-    // new_superfile_list, doesn't hydrate the freshly-written
-    // part). Contract: report what's actually in memory.
+    // Producer's in-memory state after commit: the list is set and
+    // the commit path seeds the freshly-written part into the parts
+    // cache, so it's already loaded — a same-handle query needs no
+    // refetch from storage.
     let producer_stats = producer.stats();
     assert_eq!(producer_stats.manifest_id, 1);
     assert_eq!(
@@ -112,8 +110,8 @@ fn stats_show_manifest_parts_when_storage_attached() {
         "post-commit ManifestList exists with one part"
     );
     assert_eq!(
-        producer_stats.n_manifest_parts_loaded, 0,
-        "writer's commit path doesn't hydrate the parts cache"
+        producer_stats.n_manifest_parts_loaded, 1,
+        "commit path seeds the freshly-written part into the cache"
     );
     drop(producer);
